requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
accelerate=0.27.2=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.3.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
bitsandbytes=0.42.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_5
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.2.0=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.77=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.12.1=pypi_0
dataclasses-json=0.6.4=pypi_0
dataproperty=1.1.0=pypi_0
datasets=2.18.0=pypi_0
deepspeed=0.13.5=pypi_0
dill=0.3.8=pypi_0
einops=0.7.0=pypi_0
et-xmlfile=1.1.0=pypi_0
evaluate=0.4.1=pypi_0
exceptiongroup=1.2.0=pypi_0
filelock=3.13.1=py310h06a4308_0
fonttools=4.49.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
git-filter-repo=2.38.0=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
greenlet=3.0.3=pypi_0
hjson=3.1.0=pypi_0
huggingface-hub=0.27.1=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.6=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jaxtyping=0.2.36=pypi_0
jinja2=3.1.3=py310h06a4308_0
joblib=1.3.2=pypi_0
jsonlines=4.0.0=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
kiwisolver=1.4.5=pypi_0
langchain=0.1.11=pypi_0
langchain-community=0.0.25=pypi_0
langchain-core=0.1.29=pypi_0
langchain-text-splitters=0.0.1=pypi_0
langsmith=0.1.22=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.1.6=0
libcurand=10.3.7.77=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.0.39=0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lm-eval=0.4.9=pypi_0
lxml=6.0.0=pypi_0
markupsafe=2.1.3=py310h5eee18b_0
marshmallow=3.21.1=pypi_0
matplotlib=3.8.3=pypi_0
mbstrdecoder=1.1.4=pypi_0
mkl=2023.1.0=h213fc3f_46344
more-itertools=10.7.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpi=1.0=mpich
mpi4py=3.1.4=py310hfc96bbd_0
mpich=3.3.2=hc856adb_0
mpmath=1.3.0=py310h06a4308_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=py310h06a4308_0
ninja=1.11.1.1=pypi_0
nltk=3.8.1=pypi_0
numexpr=2.11.0=pypi_0
numpy=1.26.4=pypi_0
omegaconf=2.3.0=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.13=h7f8727e_0
orjson=3.9.15=pypi_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
pathvalidate=3.3.1=pypi_0
peft=0.9.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310h06a4308_0
portalocker=3.2.0=pypi_0
psutil=5.9.8=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=15.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pybind11=3.0.0=pypi_0
pydantic=2.6.3=pypi_0
pydantic-core=2.16.3=pypi_0
pynvml=11.5.0=pypi_0
pyparsing=3.1.2=pypi_0
pytablewriter=1.2.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.2.1=py3.10_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_6
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h5eee18b_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
responses=0.18.0=pypi_0
rouge-score=0.1.2=pypi_0
sacrebleu=2.5.1=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.12.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.7=pypi_0
sqlalchemy=2.0.28=pypi_0
sqlite=3.41.2=h5eee18b_0
sqlitedict=2.1.0=pypi_0
sympy=1.12=py310h06a4308_0
tabledata=1.3.4=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tcolorpy=0.1.7=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.21.0=pypi_0
torchtriton=2.2.0=py310
torchtyping=0.1.5=pypi_0
tqdm=4.66.2=pypi_0
tqdm-multiprocess=0.0.11=pypi_0
transformers=4.47.0=pypi_0
typeguard=2.13.3=pypi_0
typepy=1.3.4=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wheel=0.41.2=py310h06a4308_0
word2number=1.1=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h5eee18b_0
zstandard=0.23.0=pypi_0
